requirements.txt
docopt==0.6.2
requests
setuptools
gitpython==3.1.12
pyyaml==5.4.1
tabulate==0.8.7
colorama==0.4.3
yaspin==1.4.1
emoji==1.2.0